+++ packages/api/src/utils.ts
@@ -1,4 +1,4 @@
-import { cors } from '@elysiajs/cors'
+import cors from '@elysiajs/cors'
 import { Elysia } from 'elysia'
 import { Logestic } from 'logestic'
 import { Cast } from './services/neynar/types'
@@ -11,34 +11,7 @@ import { getBulkPosts } from '@anonworld/db'
 
 export const createElysia = (config?: ConstructorParameters<typeof Elysia>[0]) =>
   new Elysia(config)
-    .use(
-      cors({
-        origin: (request: Request): boolean => {
-          const origin = request.headers.get('origin')
-          return (
-            origin === 'https://rumourcast.xyz' ||
-            origin === 'https://api-new.rumourcast.xyz' ||
-            origin === 'http://localhost:3000' ||
-            (origin?.endsWith('.rumourcast.xyz') ?? false)
-          )
-        },
-        methods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS', 'PATCH', 'HEAD'],
-        credentials: true,
-        allowedHeaders: [
-          'Content-Type',
-          'Authorization',
-          'Accept',
-          'Origin',
-          'X-Requested-With',
-        ],
-        exposeHeaders: [
-          'Access-Control-Allow-Origin',
-          'Access-Control-Allow-Credentials',
-        ],
-        maxAge: 86400, // 24 hours
-        preflight: true,
-      })
-    )
+    .use(cors())
     .use(Logestic.preset('common'))
     .onError(({ server, error, path }) => {
       console.error(path, error)
